// frontend/src/api/types.ts
export interface K8sObject {
  apiVersion?: string;
  kind?: string;
  dataKeys?: string[];
  metadata?: {
    name?: string;
    namespace?: string;
    creationTimestamp?: string;
    labels?: Record<string, string>;
    annotations?: Record<string, string>;
    uid?: string;
    [k: string]: unknown;
  };
  spec?: any;
  status?: any;
  data?: Record<string, string>;
  type?: string;
  [k: string]: unknown;
}